72crm_9.0.1_20190702
运行update_20190702.sql

新增：
1. 线索、客户、联系人、商机、合同跟进记录，跟进类型下拉选项自定义设置
2. 回款计划增加编辑、删除功能
3. 合同到期提前提醒天数设置
4. 公海增加导出功能

优化：
1. 权限粒度细化
2. 自定义字段处固定字段的展示
3. 客户成交状态设置为可编辑
4. 客户管理列表查看数据优化
5. 合同详情回款列表增加“回款期数”字段
6. 联系人、商机之间可以相互“关联”和“解除关联”
7. 增加在系统菜单导航栏点击鼠标右键“在新标签页中打开链接”
8. 导入导出的请求超时时间设置为1分钟
9. 客户管理-列表页点击蓝色字体才能查看对应数据详情

修复：
1. 修复新建或者编辑时自定义字段附件类型不能正常显示的问题。
2. 修复导入数据出错不能正确返回错误信息的问题。
3. 修复tomcat上传文件会找不到文件的问题
4. 修复删除数据未删除相关自定义字段的问题。

+++ src/main/java/com/kakarote/crm9/erp/oa/controller/OaEventController.java
@@ -1,9 +1,11 @@
 package com.kakarote.crm9.erp.oa.controller;
 
 import com.kakarote.crm9.common.config.paragetter.BasePageRequest;
+import com.kakarote.crm9.erp.oa.common.OaEnum;
 import com.kakarote.crm9.erp.oa.entity.OaEvent;
 import com.kakarote.crm9.erp.oa.entity.OaEventRelation;
 import com.kakarote.crm9.erp.oa.service.OaEventService;
+import com.kakarote.crm9.utils.AuthUtil;
 import com.kakarote.crm9.utils.R;
 import com.jfinal.aop.Inject;
 import com.jfinal.core.Controller;
@@ -34,6 +36,11 @@ public class OaEventController extends Controller {
      * 更新日程
      */
     public void update(@Para("")OaEvent oaEvent){
+        boolean oaAuth = AuthUtil.isOaAuth(OaEnum.EVENT_TYPE_KEY.getTypes(), oaEvent.getEventId());
+        if(oaAuth){
+            renderJson(R.noAuth());
+            return;
+        }
         renderJson(oaEventService.update(oaEvent));
     }
 
@@ -42,6 +49,11 @@ public class OaEventController extends Controller {
      * 删除日程
      */
     public void delete(@Para("eventId")Integer eventId){
+        boolean oaAuth = AuthUtil.isOaAuth(OaEnum.EVENT_TYPE_KEY.getTypes(), eventId);
+        if(oaAuth){
+            renderJson(R.noAuth());
+            return;
+        }
         renderJson(oaEventService.delete(eventId));
     }
